Time: 22 ms (72.29%), Space: 17.8 MB (93.84%) - LeetHub

+++ 268-missing-number/268-missing-number.cpp
@@ -6,7 +6,7 @@ public:
             sum+=a;
         }
         long n=nums.size();
-        long p=(n*(n+1))/2-sum;
-        return (int)p;
+        int p=(n*(n+1))/2-sum;
+        return p;
     }
 };
